User routing and controller updated for non API methods.
Using the resource method on the Route object is intended for API based
controllers.  As such, model delete and other controller methods could
not be fired.

Also, the User and Organization models are defined on the Router.  This
allows for automatic parsing of models to the controllers based on
routing params.